ISPN-9100 Add JDK-specific implementations
- infinispan-commons is now a multi-release jar
- Use StackWalker in place of sun.reflect.Reflection on JDK 10+
- Runtime.getRuntime().availableProcessors() on JDK works correctly for CGroups quotas
- use "jar" packaging for infinispan-commons instead of "bundle"
- use manifest goal of the maven-bundle-plugin
